Breakout custom vim commands into separate module.
