+++ main.js
@@ -1,54 +1,25 @@
-/*jslint vars: true, plusplus: true, devel: true, nomen: true, regexp: true, indent: 4, maxerr: 50 */
-/*global define, $, brackets, window */
-
 define(function (require, exports, module) {
     'use strict';
 
     require('autocomplete');
+    require('commands');
 
-    var AppInit         = brackets.getModule('utils/AppInit'),
-        CodeMirror      = brackets.getModule('thirdparty/CodeMirror2/lib/codemirror'),
+    var CodeMirror      = brackets.getModule('thirdparty/CodeMirror2/lib/codemirror'),
         CommandManager  = brackets.getModule('command/CommandManager'),
-        DefaultDialogs  = brackets.getModule('widgets/DefaultDialogs'),
         Dialogs         = brackets.getModule('widgets/Dialogs'),
         DocumentManager = brackets.getModule('document/DocumentManager'),
         EditorManager   = brackets.getModule('editor/EditorManager'),
         ExtensionUtils  = brackets.getModule('utils/ExtensionUtils'),
-        FileSystem      = brackets.getModule('filesystem/FileSystem'),
         KeyBindingManager = brackets.getModule('command/KeyBindingManager'),
-        MainViewManager = brackets.getModule('view/MainViewManager'),
         Menus           = brackets.getModule('command/Menus'),
         ProjectManager  = brackets.getModule('project/ProjectManager'),
         StatusBar       = brackets.getModule('widgets/StatusBar'),
-
-        utils   = require('utils'),
 
         $vimModeIndicator = $(document.createElement('div')).text('normal'),
 
         commandId   = 'brackets-vim.toggle',
         fileMenu    = Menus.getMenu(Menus.AppMenuBar.FILE_MENU),
         isEnabled   = true;
-
-    function createFile(filepath) {
-        var afterLastSlash = filepath.lastIndexOf('/') + 1,
-            filename = filepath.slice(afterLastSlash),
-            parentDir = filepath.slice(0, afterLastSlash);
-        return ProjectManager.createNewItem(parentDir, filename, true, false)
-            .fail(function () {
-                Dialogs.showModalDialog(
-                    DefaultDialogs.DIALOG_ID_ERROR,
-                    'Brackets-Vim: file creation error',
-                    'There was an error in the filepath of your vim command.  ' +
-                    'Make sure the destination directory of the file already exists.'
-                );
-            });
-    }
-
-    function doesFileExist(filepath, callback) {
-        FileSystem.resolve(filepath, function (error, item) {
-            callback(error === null);
-        });
-    }
 
     // Patch the esc key, which does not properly exit insert or visual mode in Brackets.
     // This is likely due to Brackets, since the demo at codemirror.net/demo/vim.html works fine.
@@ -68,32 +39,6 @@ define(function (require, exports, module) {
 
     function handleVimModeChange(event) {
         $vimModeIndicator.text(event.mode);
-    }
-
-    function openFile(fullPath) {
-        var currentDocument = DocumentManager.getCurrentDocument(),
-            deferred = $.Deferred();
-
-        function open() {
-            CommandManager.execute('file.addToWorkingSet', {fullPath: fullPath})
-                .done(deferred.resolve);
-        }
-
-        // Do not open the current file.
-        if (currentDocument !== null && currentDocument.file._path === fullPath) {
-            return;
-        }
-
-        doesFileExist(fullPath, function (doesExist) {
-            if (doesExist) {
-                open();
-            } else {
-                createFile(fullPath)
-                    .done(open);
-            }
-        });
-
-        return deferred;
     }
 
     // Toggles vim functionality in the editor.
@@ -128,97 +73,10 @@ define(function (require, exports, module) {
         }
     }
 
-    function closeActiveFile() {
-        return CommandManager.execute('file.close').fail(function () {
-            //TODO: implement warning in vim command bar
-        });
-    }
-    
-    function isViewSplit() {
-        return MainViewManager.getPaneCount() === 2;
-    }
-    
-    function unsplit() {
-        var layout = MainViewManager.getLayoutScheme();
-        if (layout.rows === 2) {
-            splitHorizontally();
-        }
-        if (layout.columns === 2 ) {
-            splitVertically();
-        }
-    }
-    
-    function splitHorizontally() {
-        CommandManager.execute('cmd.splitHorizontally');
-    }
-    
-    function splitVertically() {
-        CommandManager.execute('cmd.splitVertically');
-    }
-    
-    function prepareNewPane(filePath) {
-        $(MainViewManager).one('paneLayoutChange', function (jqEvent, orientation) {
-            if (orientation !== null) {
-                var file = FileSystem.getFileForPath(filePath);
-                MainViewManager.open('second-pane', file);
-            }
-        });
-    }
-    
-    // Define all custom ex commands after the vim module is loaded.
-    brackets.getModule(['thirdparty/CodeMirror2/keymap/vim'], function () {
-        CodeMirror.Vim.defineEx('edit', 'e', function (cm, params) {
-            if (!params.hasOwnProperty('args')) {
-                // The timeout is used so that the enter key can be released before the quick open dialog is shown,
-                // preventing the first file in the dialog list from being immediately opened.
-                setTimeout(function () {
-                    closeActiveFile().done(function () {
-                        CommandManager.execute('navigate.quickOpen');
-                    });
-                }, 200);
-            } else {
-                closeActiveFile().done(function () {
-                    openFile(utils.resolvePath(params.args[0]));
-                });
-            }
-        });
-
-        CodeMirror.Vim.defineEx('quit', 'q', function () {
-            closeActiveFile().done(function () {
-                if (isViewSplit()) {
-                    unsplit();
-                }
-            });
-        });
-
-        CodeMirror.Vim.defineEx('write', 'w', function () {
-            CommandManager.execute('file.save');
-        });
-
-        CodeMirror.Vim.defineEx('vsplit', 'vsp', function (cm, params) {
-            // Do nothing if already split.
-            if (isViewSplit()) {
-                return;
-            }
-
-            prepareNewPane(utils.resolvePath(params.args[0]));
-            splitVertically();
-        });
-
-        CodeMirror.Vim.defineEx('split', 'sp', function (cm, params) {
-            // Do nothing if already split.
-            if (isViewSplit()) {
-                return;
-            }
-
-            prepareNewPane(utils.resolvePath(params.args[0]));
-            splitHorizontally();
-        });
-    });
-
     // Register the enable command.
     CommandManager.register('Enable Vim', commandId, toggleVim);
     CommandManager.get(commandId).setChecked(isEnabled);
+
     // Add the enable command to the file menu.
     fileMenu.addMenuDivider();
     fileMenu.addMenuItem(commandId);
@@ -226,9 +84,6 @@ define(function (require, exports, module) {
     ExtensionUtils.loadStyleSheet(module, 'main.css');
 
     StatusBar.addIndicator('vim-mode', $vimModeIndicator, true);
-
-    // Doesn't work yet.
-    //KeyBindingManager.addBinding('navigate.nextDoc', 'Ctrl-w');
 
     $(ProjectManager).on('projectOpen', function (jqEvent, directory) {
         // Ensure that at most one file is working when the project opens.
@@ -239,8 +94,8 @@ define(function (require, exports, module) {
             return;
         }
 
-        DocumentManager.getDocumentForPath(workingSet[0]._path).done(function (document) {
-            DocumentManager.setCurrentDocument(document);
+        DocumentManager.getDocumentForPath(workingSet[0]._path).done(function (doc) {
+            DocumentManager.setCurrentDocument(doc);
         });
     });
